+++ CotEditor/Sources/UserDefaults+DefaultKey.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2016-2020 1024jp
+//  © 2016-2021 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -77,7 +77,7 @@ extension UserDefaults {
     
     subscript(key: DefaultKey<CGFloat>) -> CGFloat {
         
-        get { CGFloat(self.double(forKey: key.rawValue)) }
+        get { self.double(forKey: key.rawValue) }
         set { self.set(newValue, forKey: key.rawValue) }
     }
